@@ -1,23 +1,26 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net5.0</TargetFramework>
         <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
 
         <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
         <Description>engenious content pipeline for content compilation and similar content related things.</Description>
+        <Nullable>enable</Nullable>
     </PropertyGroup>
     <ItemGroup>
-        <PackageReference Include="Microsoft.Win32.Registry" Version="4.7.0" />
-        <PackageReference Include="Mono.Cecil" Version="0.11.2" />
-        <PackageReference Include="OpenTK" Version="4.0.0-pre9.7" />
+        <PackageReference Include="Microsoft.Win32.Registry" Version="5.0.0" />
+        <PackageReference Include="Mono.Cecil" Version="0.11.3" />
+        <PackageReference Include="MsdfGen.Net" Version="0.1.5-alpha" />
+        <PackageReference Include="OpenTK" Version="4.5.0" />
         <PackageReference Include="SharpFont.Dependencies" Version="2.6.0" PrivateAssets="all" ExcludeAssets="all" />
         <PackageReference Include="SharpFontStandard" Version="4.0.1" />
         <ProjectReference Include="..\engenious\engenious.csproj" Condition="Exists('..\engenious\engenious.csproj')" />
-        <PackageReference Include="engenious" Version="0.5.0.14-alpha" Condition="!Exists('..\engenious\engenious.csproj')" />
+        <PackageReference Include="engenious" Version="0.5.1.1-alpha" Condition="!Exists('..\engenious\engenious.csproj')" />
 
         <PackageReference Include="LitGit" Version="0.2.0.18-alpha" PrivateAssets="all" />
-        <PackageReference Include="AssimpNet" Version="4.1.0" PrivateAssets="compile" />
+        <PackageReference Include="LiCo" Version="0.1.8-alpha" PrivateAssets="all" />
+        <PackageReference Include="AssimpNet" Version="5.0.0-beta1" PrivateAssets="compile" />
     </ItemGroup>
     <ItemGroup>
         <LitGitTemplateReference Include="engenious.Pipeline.Nuget.props.template" />
