@@ -13,24 +13,22 @@ namespace engenious.Pipeline
     [ContentProcessor(DisplayName = "Model Processor")]
     public class ModelProcessor : ContentProcessor<Scene, ModelContent,ModelProcessorSettings>
     {
-        private NodeContent ParseNode(ModelContent model, Node node,NodeContent parent=null)
+        private NodeContent ParseNode(ModelContent model, Node node,NodeContent? parent = null)
         {
-            NodeContent n = new NodeContent(parent);
+            NodeContent n = new(node.Name, parent);
             model.Nodes.Add(n);
             if (settings.TransformMesh){
                 Matrix matrix = ConvertMatrix(node.Transform);
-                matrix.M41 *= settings.Scale.X;
-                matrix.M42 *= settings.Scale.Y;
-                matrix.M43 *= settings.Scale.Z;
+                matrix.M14 *= settings.Scale.X;
+                matrix.M24 *= settings.Scale.Y;
+                matrix.M34 *= settings.Scale.Z;
+
                 n.Transformation = matrix;
             }else
-                n.Transformation = Matrix.Identity;
+                n.Transformation = ConvertMatrix(node.Transform);
 
-            n.Name = node.Name;
-            n.Meshes = new List<int>();
             foreach (var meshIndex in node.MeshIndices)
                 n.Meshes.Add(meshIndex);
-            n.Children = new List<NodeContent>();
             foreach (var child in node.Children)
                 n.Children.Add(ParseNode(model, child,n));
             return n;
@@ -38,10 +36,10 @@ namespace engenious.Pipeline
 
         private Matrix ConvertMatrix(Matrix4x4 m)
         {
-            return new Matrix(m.A1, m.A2, m.A3, m.A4,
-                m.B1, m.B2, m.B3, m.B4,
-                m.C1, m.C2, m.C3, m.C4,
-                m.D1, m.D2, m.D3, m.D4);
+            return new(m.A1, m.B1, m.C1, m.D1,
+                            m.A2, m.B2, m.C2, m.D2,
+                            m.A3, m.B3, m.C3, m.D3,
+                            m.A4, m.B4, m.C4, m.D4);
         }
         private bool CombineWithParentNode(ModelContent content,NodeContent node)
         {
@@ -62,7 +60,7 @@ namespace engenious.Pipeline
             {
                 foreach(var anim in content.Animations)
                 {
-                    AnimationNodeContent c1=null,c2=null;
+                    AnimationNodeContent? c1 = null, c2 = null;
                     foreach(var c in anim.Channels)
                     {
                         if (c.Node == node.Parent)
@@ -73,7 +71,7 @@ namespace engenious.Pipeline
                         if (c1 != null && c2 != null)
                             break;
                     }
-                    CombineAnimationContent(c1,c2);
+                    CombineAnimationContent(c1, c2);
                     for(int i=anim.Channels.Count-1;i>= 0;i--)
                     {
                         if (anim.Channels[i].Node == node)
@@ -86,14 +84,14 @@ namespace engenious.Pipeline
 
             return changed;
         }
-        private void PostProcess(ModelContent content,NodeContent node)
+        private void PostProcess(ModelContent content, NodeContent node)
         {
             if (node.Name.Contains("$") && node.Name.Contains("Translation"))
                 node.Name = node.Name.Replace("Translation","Transform");
             if (node.Name.Contains("$") && !node.Name.Contains("Transform"))
             {
-                if(CombineWithParentNode(content,node))
-                    PostProcess(content,node.Parent);
+                if(CombineWithParentNode(content, node) && node.Parent != null)
+                    PostProcess(content, node.Parent);
             }
             for (int i=node.Children.Count-1;i>=0;i--){
                 var child = node.Children[i];
@@ -101,7 +99,7 @@ namespace engenious.Pipeline
                 {
                     foreach(var anim in content.Animations)
                     {
-                        AnimationNodeContent c=anim.Channels.FirstOrDefault(x=>x.Node == node);
+                        var c = anim.Channels.FirstOrDefault(x=> x.Node == node);
                         if (c == null)
                             continue;
                         foreach(var f in c.Frames)
@@ -116,34 +114,35 @@ namespace engenious.Pipeline
             }
 
         }
-        private void CombineAnimationContent(AnimationNodeContent c1,AnimationNodeContent c2)
+        private void CombineAnimationContent(AnimationNodeContent? c1, AnimationNodeContent? c2)
         {
             if (c1 == null || c2 == null)
                 return;
-            AnimationFrame empty = new AnimationFrame();
-            empty.Transform = new AnimationTransform(string.Empty,new Vector3(),new Vector3(1),new Quaternion(0,0,0,1));
+            var empty = new AnimationFrame(0f, new AnimationTransform(string.Empty,new Vector3(),new Vector3(1),new Quaternion(0,0,0,1)));
             for (int i=0;i<Math.Max(c1.Frames.Count,c2.Frames.Count);i++)
             {
-                AnimationFrame f1 = i<c1.Frames.Count?c1.Frames[i]:empty;
-                AnimationFrame f2 = i<c2.Frames.Count?c2.Frames[i]:empty;
-                if (f1== empty){
+                var f1 = i < c1.Frames.Count ? c1.Frames[i] : empty;
+                var f2 = i < c2.Frames.Count ? c2.Frames[i] : empty;
+                if (f1 == empty)
+                {
                     c1.Frames.Add(f2);
-                }else{
-                    AnimationTransform t1 = f1.Transform;
-                    AnimationTransform t2 = f2.Transform;
+                }
+                else
+                {
+                    var t1 = f1.Transform;
+                    var t2 = f2.Transform;
                     f1.Transform = t1 + t2;
                 }
             }
         }
-        public override ModelContent Process(Scene scene, string filename, ContentProcessorContext context)
+        public override ModelContent? Process(Scene scene, string filename, ContentProcessorContext context)
         {
             try
             {
                 //AssimpContext c = new AssimpContext();
                 //ExportFormatDescription des = c.GetSupportedExportFormats()[0];
                 //c.ExportFile(scene,Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Desktop),"test.dae"),des.FormatId);
-                ModelContent content = new ModelContent();
-                content.Meshes = new MeshContent[scene.MeshCount];
+                ModelContent content = new ModelContent(scene.MeshCount);
                 for (int meshIndex = 0; meshIndex < scene.MeshCount; meshIndex++)
                 {
                     var sceneMesh = scene.Meshes[meshIndex];
@@ -165,25 +164,25 @@ namespace engenious.Pipeline
                             {
                                 var pos = sceneMesh.Vertices[i];
                                 var translated = new Vector3(pos.X, pos.Y, pos.Z)+settings.Translate;
-                                meshContent.Vertices.AsPosition[vertex] =new Vector3(translated.X*settings.Scale.X,translated.Y*settings.Scale.Y,translated.Z*settings.Scale.Z);
+                                meshContent.Vertices.AsPosition![vertex] =new Vector3(translated.X*settings.Scale.X,translated.Y*settings.Scale.Y,translated.Z*settings.Scale.Z);
                             }
 
                             if (meshContent.Vertices.HasNormals)
                             {
                                 var norm = sceneMesh.Normals[i];
-                                meshContent.Vertices.AsNormal[vertex] = new Vector3(norm.X, norm.Y, norm.Z);   
+                                meshContent.Vertices.AsNormal![vertex] = new Vector3(norm.X, norm.Y, norm.Z);   
                             }
 
                             if (meshContent.Vertices.HasColors)
                             {
                                 var col = sceneMesh.VertexColorChannels[0][i];
-                                meshContent.Vertices.AsColor[vertex] = new Color(col.R,col.G,col.B,col.A);
+                                meshContent.Vertices.AsColor![vertex] = new Color(col.R,col.G,col.B,col.A);
                             }
                             if (meshContent.Vertices.HasTextureCoordinates && sceneMesh.TextureCoordinateChannels.Length > 0 && sceneMesh.TextureCoordinateChannels[0].Count > i)
                             {
                                 var tex = sceneMesh.TextureCoordinateChannels[0][i];
                                 
-                                meshContent.Vertices.AsTextureCoordinate[vertex] =new Vector2(tex.X, -tex.Y);
+                                meshContent.Vertices.AsTextureCoordinate![vertex] =new Vector2(tex.X, -tex.Y);
                             }
 
                             ++vertex;
@@ -196,43 +195,56 @@ namespace engenious.Pipeline
 
                     content.Meshes[meshIndex] = meshContent;
                 }
-                content.Nodes = new List<NodeContent>();
                 content.RootNode = ParseNode(content, scene.RootNode);
                 foreach(var animation in scene.Animations){
                     var anim = new AnimationContent();
                     anim.Channels = new List<AnimationNodeContent>();
                     foreach (var channel in animation.NodeAnimationChannels)
                     {
-                        AnimationNodeContent node = new AnimationNodeContent();
                         var curNode = content.Nodes.First(n => n.Name == channel.NodeName);
-                        node.Node = curNode;
-                        node.Frames = new List<AnimationFrame>();
+                        AnimationNodeContent node = new AnimationNodeContent(curNode);
                         int frameCount = Math.Max(Math.Max(channel.PositionKeyCount, channel.RotationKeyCount), channel.ScalingKeyCount);
                         float diff=0.0f,maxTime = 0;
                         for (int i = 0; i < frameCount; i++)
                         {
-
-                            AnimationFrame frame = new AnimationFrame();
+                            float frameTime = 0f;
 
                             if (i < channel.PositionKeyCount)
-                                frame.Frame = (float)channel.PositionKeys[i].Time;
+                                frameTime = (float)channel.PositionKeys[i].Time;
                             else if (i < channel.RotationKeyCount)
-                                frame.Frame = (float)channel.RotationKeys[i].Time;
+                                frameTime = (float)channel.RotationKeys[i].Time;
                             else if (i < channel.ScalingKeyCount)
-                                frame.Frame = (float)channel.ScalingKeys[i].Time;
-                            if (i==0)
-                                diff = frame.Frame;
-                            frame.Frame -= diff;
-                            frame.Frame = (float)(frame.Frame / animation.TicksPerSecond);
-                            maxTime = Math.Max(frame.Frame, maxTime);
+                                frameTime = (float)channel.ScalingKeys[i].Time;
+                            if (i == 0)
+                                diff = frameTime;
+                            frameTime -= diff;
+                            frameTime = (float)(frameTime / animation.TicksPerSecond);
+                            maxTime = Math.Max(frameTime, maxTime);
                             //TODO: interpolation
                             var rot = channel.RotationKeyCount == 0 ? new Assimp.Quaternion(1, 0, 0, 0) : i >= channel.RotationKeyCount ? channel.RotationKeys.Last().Value : channel.RotationKeys[i].Value;
                             var loc = channel.PositionKeyCount == 0 ? new Vector3D() : i >= channel.PositionKeyCount ? channel.PositionKeys.Last().Value : channel.PositionKeys[i].Value;
                             var sca = channel.ScalingKeyCount == 0 ? new Vector3D(1, 1, 1) : i >= channel.ScalingKeyCount ? channel.ScalingKeys.Last().Value : channel.ScalingKeys[i].Value;
                             rot.Normalize();
+                            
+                            var relativeTransform = Matrix.Invert(node.Node.Transformation);
+                            
+                            var transform = new AnimationTransform(node.Node.Name,
+                                new Vector3((loc.X + settings.Translate.X), (loc.Y + settings.Translate.Y),
+                                    (loc.Z + settings.Translate.Z)),
+                                new Vector3(sca.X * settings.Scale.X, sca.Y * settings.Scale.Y,
+                                    sca.Z * settings.Scale.Z), new Quaternion(rot.X, rot.Y, rot.Z, rot.W));
 
-                            frame.Transform = new AnimationTransform(node.Node.Name,new Vector3((loc.X+settings.Translate.X), (loc.Y+settings.Translate.Y), (loc.Z+settings.Translate.Z)),
-                                new Vector3(sca.X*settings.Scale.X, sca.Y*settings.Scale.Y, sca.Z*settings.Scale.Z), new Quaternion(rot.X, rot.Y, rot.Z, rot.W));
+                            var tmp = transform.ToMatrix() * relativeTransform;
+
+                            var (relLoc, relScal, relRot) = tmp;
+
+                            transform = new AnimationTransform(node.Node.Name,
+                                relLoc, relScal, relRot);
+
+                            // if (tmp != transform.ToMatrix())
+                            //     throw new Exception();
+                            
+                            AnimationFrame frame = new AnimationFrame(frameTime, transform);
                             node.Frames.Add(frame);
                         }
                         anim.MaxTime = maxTime;
