@@ -5,7 +5,7 @@ namespace engenious.Content.Pipeline
     [AttributeUsage(AttributeTargets.Class)]
     public class ContentProcessorAttribute : Attribute
     {
-        public virtual string DisplayName{ get; set; }
+        public virtual string? DisplayName{ get; set; }
     }
 }
 
