@@ -7,9 +7,9 @@ namespace engenious.Content.Pipeline
         public Type ExportType => _exportType;
         protected static readonly Type _exportType = typeof (T);
 
-        public abstract T Import(string filename, ContentImporterContext context);
+        public abstract T? Import(string filename, ContentImporterContext context);
 
-        object IContentImporter.Import(string filename, ContentImporterContext context)
+        object? IContentImporter.Import(string filename, ContentImporterContext context)
         {
             return Import(filename, context);
         }
