@@ -5,32 +5,26 @@ namespace engenious.Pipeline
 {
     public abstract class FontConfig
     {
-        private static FontConfig _fontConfig;
+        private static FontConfig? _fontConfig;
 
         public static FontConfig Instance
         {
             get
             {
-                if (_fontConfig != null) return _fontConfig;
-                switch (PlatformHelper.RunningPlatform())
+                if (_fontConfig != null)
+                    return _fontConfig;
+                _fontConfig = PlatformHelper.RunningPlatform() switch
                 {
-                    case Platform.Linux:
-                        _fontConfig = new FontConfigLinux();
-                        break;
-                    case Platform.Mac:
-                        _fontConfig = new FontConfigMac();
-                        break;
-                    case Platform.Windows:
-                        _fontConfig = new FontConfigWindows();
-                        break;
-                    default:
-                        throw new NotSupportedException();
-                }
+                    Platform.Linux => new FontConfigLinux(),
+                    Platform.Mac => new FontConfigMac(),
+                    Platform.Windows => new FontConfigWindows(),
+                    _ => throw new NotSupportedException()
+                };
                 return _fontConfig;
             }
         }
 
         public abstract bool GetFontFile(string fontName, int fontSize, System.Drawing.FontStyle style,
-            out string fileName);
+            out string? fileName);
     }
 }
