@@ -6,7 +6,7 @@ namespace engenious.Content.Pipeline
     [ContentImporter(".bmp", ".jpg", ".png", DisplayName = "Bitmap Importer", DefaultProcessor = "BitmapProcessor")]
     public class BitmapImporter : ContentImporter<Bitmap>
     {
-        public override Bitmap Import(string filename, ContentImporterContext context)
+        public override Bitmap? Import(string filename, ContentImporterContext context)
         {
             //if (!System.IO.File.Exists(filename))
             //    return null;
