@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using engenious.Content;
 using engenious.Content.Pipeline;
 using engenious.Graphics;
@@ -10,11 +11,14 @@ namespace engenious.Pipeline
     {
         #region implemented abstract members of ContentImporter
 
-        public override ModelContent Import(string filename, ContentImporterContext context)
+        public override ModelContent? Import(string filename, ContentImporterContext context)
         {
-
-            ContentManager manager = new ContentManager(null,Path.GetDirectoryName(filename));
-            return manager.Load<ModelContent>(Path.GetFileNameWithoutExtension(filename));
+            var dirName = Path.GetDirectoryName(filename) ??
+                          throw new DirectoryNotFoundException($"Could not get directory name of {filename}]");
+            var fileWithoutExt = Path.GetFileNameWithoutExtension(filename) ??
+                                 throw new ArgumentException("Not a valid filename", nameof(filename));
+            ContentManagerBase managerBase = new AggregateContentManager(null!, dirName);
+            return managerBase.Load<ModelContent>(fileWithoutExt);
         }
 
         #endregion
