@@ -7,17 +7,17 @@ using engenious.Helper;
 
 namespace engenious.Pipeline
 {
-    [ContentImporter(".3D",".3DS",".3MF",".AC",".AC3D",".ACC",".AMJ",".ASE",".ASK",".B3D",".BLEND",".BVH",".COB",".CMS",".DAE",".DXF",".ENFF",".FBX",".glTF",".glTF",".HMB",".IFC-STEP",".IRR",".LWO",".LWS",".LXO",".MD2",".MD3",".MD5",".MDC",".MDL",".MESH",".MOT",".MS3D",".NDO",".NFF",".OBJ",".OFF",".OGEX",".PLY",".PMX",".PRJ",".Q3O",".Q3S",".RAW",".SCN",".SIB",".SMD",".STL",".STP",".TER",".UC",".VTA",".X",".X3D",".XGL",".ZGL", DisplayName = "Model Importer", DefaultProcessor = "ModelProcessor")]
+    [ContentImporter(".3D",".3DS",".3MF",".AC",".AC3D",".ACC",".AMJ",".ASE",".ASK",".B3D",".BLEND",".BVH",".COB",".CMS",".DAE",".DXF",".ENFF",".FBX",".glTF",".glTF", ".glB",".HMB",".IFC-STEP",".IRR",".LWO",".LWS",".LXO",".MD2",".MD3",".MD5",".MDC",".MDL",".MESH",".MOT",".MS3D",".NDO",".NFF",".OBJ",".OFF",".OGEX",".PLY",".PMX",".PRJ",".Q3O",".Q3S",".RAW",".SCN",".SIB",".SMD",".STL",".STP",".TER",".UC",".VTA",".X",".X3D",".XGL",".ZGL", DisplayName = "Model Importer", DefaultProcessor = "ModelProcessor")]
     public class ModelImporter : ContentImporter<Scene>
     {
 
-        private static readonly Exception DllLoadExc;
+        private static readonly Exception? DllLoadExc;
 
         static ModelImporter()
         {
             try
             {
-                string dir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+                var dir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
                 if (dir == null)
                     throw new Exception("executing path not found");
                 string ext = ".dll";
@@ -48,7 +48,7 @@ namespace engenious.Pipeline
             }
         }
 
-        public override Scene Import(string filename, ContentImporterContext context)
+        public override Scene? Import(string filename, ContentImporterContext context)
         {
             if (DllLoadExc != null)
                 context.RaiseBuildMessage("FBXIMPORT" , DllLoadExc.Message, BuildMessageEventArgs.BuildMessageType.Error);
