@@ -4,12 +4,19 @@ namespace engenious.Content.Serialization
 {
 	public abstract class ContentTypeWriter<T> : IContentTypeWriter
 	{
-	    public abstract void Write (ContentWriter writer, T value);
-
-		void IContentTypeWriter.Write (ContentWriter writer, object value)
+		protected ContentTypeWriter(uint contentVersion)
 		{
-			Write (writer, (T)value);
+			ContentVersion = contentVersion;
 		}
+
+		public abstract void Write (ContentWriter writer, T? value);
+
+		void IContentTypeWriter.Write (ContentWriter writer, object? value)
+		{
+			Write (writer, (T?)value);
+		}
+
+		public uint ContentVersion { get; }
 
 		public abstract string RuntimeReaderName{ get; }
 
