@@ -4,11 +4,13 @@ namespace engenious.Content.Serialization
 {
 	public interface IContentTypeWriter
 	{
-		void Write (ContentWriter writer, object value);
+		void Write (ContentWriter writer, object? value);
 
 		string RuntimeReaderName{ get; }
 
 		Type RuntimeType{ get; }
+		
+		public uint ContentVersion { get; }
 	}
 }
 
