@@ -7,7 +7,7 @@ namespace engenious.Content.Serialization
     {
         #region implemented abstract members of ContentTypeWriter
 
-        public override void Write(ContentWriter writer, RasterizerState value)
+        public override void Write(ContentWriter writer, RasterizerState? value)
         {
             if (value == null)
             {
@@ -25,17 +25,21 @@ namespace engenious.Content.Serialization
             writer.Write(value.SlopeScaleDepthBias);
         }
 
-        public override string RuntimeReaderName => typeof(RasterizerStateTypeReader).FullName;
+        public override string RuntimeReaderName => typeof(RasterizerStateTypeReader).FullName!;
 
         #endregion
+
+        public RasterizerStateTypeWriter() : base(0)
+        {
+        }
     }
 
     [ContentTypeWriter]
-    public class DepthStencilStateTypeWriter  : ContentTypeWriter<DepthStencilState>
+    public class DepthStencilStateTypeWriter : ContentTypeWriter<DepthStencilState>
     {
         #region implemented abstract members of ContentTypeWriter
 
-        public override void Write(ContentWriter writer, DepthStencilState value)
+        public override void Write(ContentWriter writer, DepthStencilState? value)
         {
             if (value == null)
             {
@@ -63,9 +67,13 @@ namespace engenious.Content.Serialization
             writer.Write((ushort)value.CounterClockwiseStencilPass);
         }
 
-        public override string RuntimeReaderName => typeof(DepthStencilStateTypeReader).FullName;
+        public override string RuntimeReaderName => typeof(DepthStencilStateTypeReader).FullName!;
 
         #endregion
+
+        public DepthStencilStateTypeWriter() : base(0)
+        {
+        }
     }
 
     [ContentTypeWriter]
@@ -73,7 +81,7 @@ namespace engenious.Content.Serialization
     {
         #region implemented abstract members of ContentTypeWriter
 
-        public override void Write(ContentWriter writer, BlendState value)
+        public override void Write(ContentWriter writer, BlendState? value)
         {
             if (value == null)
             {
@@ -95,9 +103,13 @@ namespace engenious.Content.Serialization
             writer.Write((byte)value.ColorWriteChannels3);
         }
 
-        public override string RuntimeReaderName => typeof(BlendStateTypeReader).FullName;
+        public override string RuntimeReaderName => typeof(BlendStateTypeReader).FullName!;
 
         #endregion
+
+        public BlendStateTypeWriter() : base(0)
+        {
+        }
     }
 }
 
