@@ -1,4 +1,5 @@
-﻿using engenious.Content.Pipeline;
+﻿using System;
+using engenious.Content.Pipeline;
 
 namespace engenious.Content.Serialization
 {
@@ -7,8 +8,10 @@ namespace engenious.Content.Serialization
     {
         #region implemented abstract members of ContentTypeWriter
 
-        public override void Write(ContentWriter writer, TextureContent value)
+        public override void Write(ContentWriter writer, TextureContent? value)
         {
+            if (value == null)
+                throw new ArgumentNullException(nameof(value), "Cannot write null TextureContent");
             writer.Write(value.GenerateMipMaps);
             writer.Write(value.MipMapCount);
             foreach(var map in value.MipMaps)
@@ -17,9 +20,13 @@ namespace engenious.Content.Serialization
             }
         }
 
-        public override string RuntimeReaderName => typeof(Texture2DTypeReader).FullName;
+        public override string RuntimeReaderName => typeof(Texture2DTypeReader).FullName!;
 
         #endregion
+
+        public TextureContentTypeWriter() : base(0)
+        {
+        }
     }
 }
 
